airflow/requirements.txt
apache-airflow==2.5.1
gitpython==3.1.31